user create with person

+++ db/migrations/createListPeopleTable.js
@@ -4,7 +4,7 @@ module.exports = {
       'listPeople',
       {
         id: {
-          type: Sequelize.INTEGER,
+          type: Sequelize.BIGINT,
           primaryKey: true,
           autoIncrement: true
         },
@@ -16,12 +16,12 @@ module.exports = {
         },
 
         'teamId': {
-          type: Sequelize.INTEGER,
+          type: Sequelize.BIGINT,
           allowNull: false
         },
 
         'listId': {
-          type: Sequelize.INTEGER,
+          type: Sequelize.BIGINT,
           allowNull: false
         },
         'personGuid': {
